feat: add mob skill WZ data and object timers
- Load MobSkill.img.xml with skill slots, attacks, and banish
- Generalize map timers from characters to all map objects
- Fix mob buff and MoveMob protocol fields for skill level and action
- Add ApplyMobBuffReflect response and rect helpers for skill bounds

+++ services/game/server/map_listener.go
@@ -310,7 +310,7 @@ func (l *MapListenerImpl) OnMobMobBuffApplied(mapInstance *entity.Map, mob *enti
 		OID:        mob.OID,
 		Status:     int32(buff),
 		X:          int16(value),
-		SkillID:    skillID,
+		SkillID:    uint16(skillID),
 		BuffTime:   buffTime,
 		Delay:      0,
 		StatusSize: 1,
